make getCustomersAmount return long
change var totalAmountInDataBase in ListSubset to long

+++ Backend/src/main/server/sql/dto/ListSubset.java
@@ -7,10 +7,10 @@ import java.util.List;
 @Data
 public class ListSubset<T> {
 	private final List<T> listSubset;
-	private final int totalAmountInDataBase;
+	private final long totalAmountInDataBase;
 
 
-	public ListSubset(List<T> subsetList, int totalAmountInDataBase) {
+	public ListSubset(List<T> subsetList, long totalAmountInDataBase) {
 		this.listSubset = subsetList;
 		this.totalAmountInDataBase = totalAmountInDataBase;
 	}
